get_healthcheck_for_instance a member of long_running_service_tools
