+++ paasta_tools/cli/cmds/local_run.py
@@ -38,8 +38,8 @@ from paasta_tools.cli.utils import get_instance_config
 from paasta_tools.cli.utils import lazy_choices_completer
 from paasta_tools.cli.utils import list_instances
 from paasta_tools.cli.utils import list_services
+from paasta_tools.long_running_service_tools import get_healthcheck_for_instance
 from paasta_tools.marathon_tools import CONTAINER_PORT
-from paasta_tools.marathon_tools import get_healthcheck_for_instance
 from paasta_tools.paasta_execute_docker_command import execute_in_container
 from paasta_tools.utils import _run
 from paasta_tools.utils import DEFAULT_SOA_DIR
